requirements.txt
requests
pandas
websocket-client
matplotlib
pillow
urllib3
pydantic
plotly